assignments/cs4395_portfolio_chatbot/chatbot_code/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
bs4 = "^0.0.1"
nltk = "^3.7"
requests = "^2.28.1"
chatterbot-voice = "^0.1.5"
spacy = "^3.4.3"
PyYAML = "^6.0"
python-Levenshtein = "^0.20.8"
spacytextblob = "^4.0.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"